Add files via upload
Hi Tim,

I was able to fix only the file error when user tries to view the credentials first before adding.  I am not able to fix the format errors.  I have other commitments to attend to now so I am ok to accept the grading accordingly, hoping I can still pass.  But if required to be fixed, then I have to do it on weekend.

Thank you.